[step2] fix: 이익률 값 잘못 나오는 부분 수정

+++ src/test/kotlin/lotto/MyLottoTicketsTest.kt
@@ -33,6 +33,6 @@ class MyLottoTicketsTest {
 
         val myProfilt = myLottoTickets.getProfit(MyLottoResult(mapOf(LottoWinnerRank.FOURTH_PRICE to 1)))
 
-        assertThat(myProfilt).isEqualTo((5000 / 3 * 1000).toDouble())
+        assertThat(myProfilt).isEqualTo(((5000).toDouble() / (3 * 1000).toDouble()))
     }
 }
